requirements.txt
pytest==6.2.5
selenium==3.141.0
webdriver-manager
allure-pytest
pytest-xdist
faker